initial code payload
Okay so this is a large commit, but I went until I could get everything
to typecheck. Essentially, I added modules for project types including
`User` and `APIQuery` as well as a refactor-ready module for SQL queries
and finally an initial resource declaration for the user. I also wrote
a few typeclass instances for the UUID type since I like UUIDs